Added "--verbose", fixed 'changed', file cleanup
I really needed a “—verbose” flag for day to day usage and to ease the
initial setup right for something like a Demandware WebDAV server. It
was hard to know basics like whether or not a file was successfully
deployed, so now that happens with timestamps.

I modified the ‘changed’ state so that it’s more vigilant. Things like
byte-level changes to sass sourcemaps were getting ignored because
atime > mtime. This was happening to me a lot and became a serious
pain, so I’ve removed that chunk of code.

modified the error and stdout states, which meant modifying the default
curl commands, you can see this in cli.js

Linted, headers a lil more commented, and coffeescripts removed. Also a
few stylistic changes with the text and little circle/square icons for
upload status.